requirements.txt
graphql-core==3.0.5
Jinja2==2.11.2
